+++ NEWS.md
@@ -3,7 +3,8 @@
 * Fix internal -log10p calculation
 * Swap most "Depends" to "Imports"
 * Add dependency for {ieugwasr}
-* Add `ld_clump` option to get independent SNPs at the same locus
+* Add `ld_clump` option to `get_loci()` function, to get independent SNPs at the same locus
+* Add `use_pvalue` option to `get_loci()` function
 
 # gwasRtools 0.1.0